# contracts/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WeChat - 最新版本 iPad859 - Android858 - win4.0.3.43 - IOS18.3.1（16 Pro Max）- Android-34(HUAWEI Mate XT)-（a16+62） - 五端登陆",
    "description": "各种真机可选，设备名称：1、iPad Pro 12.9-inch (M4) 2、iPad Pro 11-inch (M4) 3、iPad Pro 12.9-inch (M3) 4、iPad Pro 11-inch (M3) 5、iPad Pro 12.9-inch (M2) 6、iPad Pro 11-inch (M2) 7、iPad Air (M2) 8、iPad Air (M1) 9、iPad 10th generation 10、iPad mini (A15) 11、HUAWEI MatePad Pro 12、Xiaomi Pad 6 Pro 13、Samsung Galaxy Tab S9 Ultra 14、Windows 11 PC 15、Windows 10 PC 16、Windows 8 PC 17、Windows 7 PC 18、车载设备",
    "version": "2.0"
  },
  "basePath": "/api",
  "paths": {
...
    "/Login/GetQR": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "获取二维码(iPad) - 8.0.59",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "不使用代理请留空",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login.GetQRReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
},
  "definitions": {
...
    "Login.GetQRReq": {
      "title": "GetQRReq",
      "type": "object",
      "properties": {
        "DeviceID": {
          "type": "string"
        },
        "DeviceName": {
          "type": "string"
        },
        "Proxy": {
          "$ref": "#/definitions/models.ProxyInfo"
        },
        "Type": {
          "type": "string",
          "description": "设备类型ID，用于选择设备配置。1-10为iPad设备，11-13为安卓设备，14-17为Windows设备，18为车载设备，其他值用于Mac设备"
        }
      }
    },
...
    "models.ProxyInfo": {
      "title": "ProxyInfo",
      "type": "object",
      "properties": {
        "ProxyIp": {
          "type": "string"
        },
        "ProxyPassword": {
          "type": "string"
        },
        "ProxyUser": {
          "type": "string"
        }
      }
    }
  },
...
}